feat(app): update app, use new modules
