+++ README.md
@@ -9,3 +9,8 @@
 - `pip install -r requierments.txt`
 
 - `streamlit run app.py`
+
+## Notes
+
+OpenStreetMap request method needs to be updated following the latest version of osmnx.
+See this [issue](https://github.com/gboeing/osmnx/issues/832)
